refactor(media): normalize html and react entrypoints
