+++ packages/html/src/media/mux-audio/index.ts
@@ -1,26 +1,1 @@
-import { CustomMediaElement } from '@videojs/media/dom/custom-media-element';
-import { GoogleCast } from '@videojs/media/dom/google-cast';
-import { addMediaComponent } from '@videojs/media/dom/media-host';
-import { MuxData, MuxMedia } from '@videojs/media/dom/mux';
-import { MediaAttachMixin } from '../../store/media-attach-mixin';
-
-const MuxAudioBase = MediaAttachMixin(CustomMediaElement('audio', MuxMedia));
-
-export class MuxAudio extends MuxAudioBase {
-  constructor() {
-    super();
-    addMediaComponent(this.host, new MuxData({ playerSoftwareName: 'mux-audio' }));
-    addMediaComponent(this.host, new GoogleCast());
-    this.host.addEventListener('sourcechange', () => this.#reflectSrc());
-  }
-
-  // Mirrors the host `src` to the `src` attribute so it matches the active playback URL.
-  #reflectSrc() {
-    const src = this.host.src;
-    if (src) {
-      if (this.getAttribute('src') !== src) this.setAttribute('src', src);
-    } else if (this.hasAttribute('src')) {
-      this.removeAttribute('src');
-    }
-  }
-}
+export * from './media';
