+++ packages/html/src/media/simple-hls-video/index.ts
@@ -1,5 +1,1 @@
-import { CustomMediaElement } from '@videojs/media/dom/custom-media-element';
-import { SimpleHlsMedia } from '@videojs/media/dom/simple-hls';
-import { MediaAttachMixin } from '../../store/media-attach-mixin';
-
-export class SimpleHlsVideo extends MediaAttachMixin(CustomMediaElement('video', SimpleHlsMedia)) {}
+export * from './media';
