+++ site/scripts/api-docs-builder/src/tests/fixtures/monorepo/packages/html/src/media/embed-video/media.ts
@@ -1,0 +1,10 @@
+import { CustomMediaElement } from '../../../../media/src/dom/custom-media-element';
+import { EmbedHost } from '../../../../media/src/dom/embed';
+
+function MediaAttachMixin(base: any) {
+  return base;
+}
+
+class EmbedCustomMediaElement extends CustomMediaElement('iframe', EmbedHost) {}
+
+export class EmbedVideo extends MediaAttachMixin(EmbedCustomMediaElement) {}
